This is a version v0.3.6, which has been tested with BioProject data.
I need to revise the read-selection part because I realized that
that inter-contig selection might not work as I expected.
We need to consider the following:
1. use ptGAUL selection method
2. use single-mapped only
3. use single- and pari-mapped

Additionally, we might want to use long-read data not sampled to select
reads. So, we would have long-read data file:
0. l.fq or l.fq.gz: long-read data file
1. n3k.fq.gz: sampled by the rate determined with estimated genome size
   used for the whole-genome assembly
2. l3k.fq.gz: long-read data cut by 3 kb or not sampled

We will test a range of the single-mapped and pair-mapped minimum. We
may need to test the ratio 0.7 used in read-selection. We will use the
total number bases or number of fragments as a metric instead of
sequence pairwise alignment or sequence identity.